requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
python-dotenv
requests
langchain==1.0.0
langchain-google-genai
langgraph-prebuilt==1.0.4
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2